requirements.txt
pytest==8.3.4
playwright==1.42.0
pytest-playwright==0.6.2
python-dotenv==1.0.0  # For environment variables